# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.2.0",
    "description": "Go-native Rebecca panel API. This document replaces the former FastAPI /docs page and is intended for dashboard, bot, and third-party client developers.\n\nAuthentication: protected routes use `Authorization: Bearer <token>`. The token can be the JWT returned by `POST /api/admin/token`; admin API keys are also accepted by the Go auth layer where bearer validation is used. Public subscription and home/docs routes explicitly set `security: []`.\n\nResponse compatibility: most endpoints intentionally keep the JSON shape expected by the current dashboard. Some schemas are documented as generic JSON while the migration to strict OpenAPI component schemas continues.\n\nMutation behavior: user, service, node, config, and host mutations are designed to be transactional. When runtime sync is required, Rebecca writes `node_operations` in the same transaction so failed queue writes roll back the database change."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Authenticate admin",
        "description": "Authenticate an admin and issue a JWT access token. This is the compatibility alias used by the former FastAPI OAuth2 form route.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: accepts `application/x-www-form-urlencoded` for OAuth2-compatible clients and JSON for programmatic clients.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/api/admin/token": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Authenticate admin",
        "description": "Authenticate an admin and issue a JWT access token.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: accepts `application/x-www-form-urlencoded` for OAuth2-compatible clients and JSON for programmatic clients.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/LoginForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Token"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdminToken"
      }
    },
    "/internal/admin/validate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Validate admin token",
        "description": "Internal validation endpoint for JWT and admin API key authentication.\n\nAuthentication: public route; no bearer token is required.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "security": [],
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postInternalAdminValidate"
      }
    },
    "/api/admin": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get current admin",
        "description": "Return the authenticated admin context, role, permissions, limits, and service assignments.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmin"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create admin",
        "description": "Create a new admin if the current admin can manage admins.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postAdmin"
      }
    },
    "/api/admins": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List admins",
        "description": "List non-deleted admins with optional username filtering, pagination, and sorting.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getAdmins"
      }
    },
...
    "/api/admin/permissions/standard/bulk": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Bulk standard-admin permissions",
        "description": "Disable or restore selected standard-admin user permissions in bulk.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Admin routes require an authenticated bearer token except token creation and internal validation. Mutations generally require admin-management permission or sudo/full_access; full_access role changes require a full_access actor.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postAdminPermissionsStandardBulk"
      }
    },
...
    "/api/myaccount/change_password": {
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Change password",
        "description": "Change the current admin password and invalidate older tokens.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires any active authenticated admin. The target account is always the current bearer principal.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postMyaccountChangePassword"
      }
    },
    "/api/myaccount/api-keys": {
      "get": {
        "tags": [
          "My Account"
        ],
        "summary": "List personal API keys",
        "description": "List API keys owned by the current admin.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires any active authenticated admin. The target account is always the current bearer principal.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getMyaccountKeys"
      },
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Create personal API key",
        "description": "Create a new admin API key for the current account.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires any active authenticated admin. The target account is always the current bearer principal.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postMyaccountKeys"
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "List users with search, filters, sorting, owner/service scope, optional links, and traffic visibility enforcement.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRoute details: Use repeated `username` or `filter` query parameters when needed. `links=true` asks the server to include subscription/config link fields, which costs extra DB/template work.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "service_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "links",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getUsers"
      }
    },
    "/api/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Create a user with required service assignment, permission checks, admin/service limits, credential normalization, and node operation enqueue.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Required business fields: `username` and `service_id`. `service_id` must point to an existing non-default service. Manual `inbounds` and `proxies` payloads are rejected; `credential_key` may be omitted, a Rebecca key, or a UUID that is normalized into the stored credential key.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/JSON"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postUser"
      }
    },
    "/api/v2/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Versioned alias for creating a user.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: This is the v2 alias of `/api/user`; it has the same validation and service-required behavior.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postV2Users"
      }
    },
...
    "/api/users/actions": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Bulk user action",
        "description": "Run bulk actions such as expire adjustment, traffic adjustment, status cleanup, activation, disable, and service change.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Supported actions include expire adjustments, traffic adjustments, cleanup_status, activate_users, disable_users, and change_service. Service/admin scopes and per-service limits are enforced server-side.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postUsersActions"
      }
    },
    "/api/users/usage": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Users usage analytics",
        "description": "Return usage analytics grouped by admin for visible users.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an active authenticated admin. Standard admins are limited to users they own and services assigned to them; traffic fields may be zeroed when visibility is disabled.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getUsersUsage"
      }
    },
    "/api/v2/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List services",
        "description": "List services visible to the current admin with usage, host, user, and broken-service summary.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin. Listing/detail respects service visibility. Service mutation, admin limits, reset, and auto-inbound actions require sudo/full_access.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getV2Services"
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Create service",
        "description": "Create a service, assign hosts, assign admins, and persist all changes transactionally.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin. Listing/detail respects service visibility. Service mutation, admin limits, reset, and auto-inbound actions require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postV2Services"
      }
    },
...
    "/api/nodes/usage": {
      "get": {
        "tags": [
          "Nodes"
        ],
        "summary": "Nodes usage",
        "description": "Return aggregate node usage over a date range.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires sudo/full_access. Node calls use Go node controller and gRPC/mTLS where runtime communication is needed.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Start"
          },
          {
            "$ref": "#/components/parameters/End"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getNodesUsage"
      }
    },
...
    "/api/node": {
      "post": {
        "tags": [
          "Nodes"
        ],
        "summary": "Create node",
        "description": "Create a node using generated or pending certificate material and enqueue initial sync.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires sudo/full_access. Node calls use Go node controller and gRPC/mTLS where runtime communication is needed.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Create payload may include certificate material directly or a `certificate_token` returned by `/api/node/certificate/new`. The response includes `node_certificate` and one-time `node_certificate_key`; the dashboard combines them into a single node install bundle for the installer. `add_as_new_host` is removed and ignored by the Go flow.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postNode"
      }
    },
...
    "/api/core/restart": {
      "post": {
        "tags": [
          "Core"
        ],
        "summary": "Queue runtime restart",
        "description": "Queue sync/restart behavior for the selected config target without starting local Xray on the master.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires sudo/full_access for config/runtime/geo/WARP/outbound helper mutations. `/api/core` and `/api/core/ips` can be read by authenticated admins.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "master"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postCoreRestart"
      }
    },
...
    "/api/core/ips": {
      "get": {
        "tags": [
          "Core"
        ],
        "summary": "Get public IPs",
        "description": "Return public IPv4/IPv6 for the master or for a node target.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires sudo/full_access for config/runtime/geo/WARP/outbound helper mutations. `/api/core` and `/api/core/ips` can be read by authenticated admins.",
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getCoreIps"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or Rebecca admin API key"
      }
    },
...
    "responses": {
...
      "Token": {
        "description": "Admin access token.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                },
                "token_type": {
                  "type": "string",
                  "default": "bearer"
                }
              }
            }
          }
        }
      },
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
...
  }
}